feat: include username in mention logic
